Construction of block cipher SPECK and reduced round attack on the Cipher
This block cipher was published by the U.S. National Security Agency in
2013 paper. An efficient implementation of this cipher was accomplished
using bit manipulation technique. To this cipher an attack up to 3 round
was also devised. The principal of this algorithm was based on the meet
in the middle approach. To improve the efficiency, CPU profiling was
considered in implementation.